add velocity control that works correctly

+++ RobotCode/PIDLoop.h
@@ -4,7 +4,7 @@
 class PIDLoop {
   public:
     PIDLoop(float kP1, float kI1, float kD1, float kF1);
-    float updateError(float target, float actual, unsigned long dt);
+    float updateError(float target, float actual, double dt);
     void wipe();
 
   private:
